Moved lollipop series to subfolder
