+++ js/masters/modules/lollipop.src.js
@@ -8,4 +8,4 @@
  * License: www.highcharts.com/license
  */
 'use strict';
-import '../../Series/LollipopSeries.js';
+import '../../Series/Lollipop/LollipopSeries.js';
